Create BaseRule service
This will be the entry point of our application for calculating discounts.

The plan is to have this as the base class for the rest of the expected possible
discounts. These classes should inherit from the base model.

For now it will always expect to have a base price, and it is expected for
the extending classes to implement the calculate method.