Introduce BulkWithNewPrice rule
This commit introduces a new rule to our set. It will require new fields
on the initialization, to be aware of what is the expected new price
that the product should have, and what's the threshold to apply this change.

This leaves our interface for the initialization a bit different, but
we remain with the same calculate interface.

Potentially starting to see a pattern, where the rules are checking that
the quantity needs to be >= 0, so maybe in the future it is worth to move
that logic to the BaseRule instead.