Introduce the BulkWithDiscount rule
This commit introduces the final rule from our requirement list, where
we want to be able to change the product final price if they meet the
threshold criteria.

Similar to the previous rule, we are changing only the initialize interface
and leaving the calculate the same.

For this one we are introducing an extra check on the discount quantity to ensure
we don't possibly end up with "free" products or instead of applying a discount
we end up applying an overcharge (Assuming that we don't want that).

It falls out of scope of the test requirement, but potentially this field
could also allow receiving a string and veryfing if it is a percentage, so that
we can handle a wider variety instead of having fractions. This last part is
important, since a user introducing `1/2` without `.to_f` would return `0`.