Create BasketTotalCalculator class
This new class should be the one responsible for calculating the total
final price. It will return an array containing the total final price
and an extra array showing any product that is not available on our
product list yet. These not found products will not be considered for
the pricing calculation.

A couple of things could be improved. I have hardcoded the mapping for
which product goes against which rule. I think this is best handled in
database. Also, I encountered a problem with ruby loading paths where
the calculator service does not have the rules loaded yet returning
undefined constant errors.