Add executable that starts a small CLI to receive a basket
This commit introduces a bin/amenitiz file that when called will simply
start a small CLI, consiting of an infinite loop that requests the user
for a list of basket items.

This will run the provided list through our calculator and print a formatted
table to the user showing the received list, the expected total, and if there
were any products that are not available on our list.